Move `DEFAULT_GRADIENT`, `GRADIENT` and `GRADIENT_REGEX` to `constants.ts`
