+++ packages/patterns/src/index.ts
@@ -15,6 +15,7 @@
  */
 
 export * from './types';
+export * from './constants';
 
 export { default as convertToCSS } from './convertToCSS';
 export { default as createSolid } from './createSolid';
@@ -31,7 +32,4 @@ export { default as hasOpacity } from './hasOpacity';
 export { default as isPatternEqual } from './isPatternEqual';
 export { default as isHexColorString } from './isHexColorString';
 export { default as getGradientStyleFromColor } from './getGradientStyleFromColor';
-export {
-  default as getColorFromGradientStyle,
-  DEFAULT_GRADIENT,
-} from './getColorFromGradientStyle';
+export { default as getColorFromGradientStyle } from './getColorFromGradientStyle';
